+++ CHANGELOG.rst
@@ -17,6 +17,7 @@ Added
 =====
 - When a link is added and one of the interfaces already has a link a warning would be log to indicate a mismatched link.
 - When two links shared an interface, the last added link will not be unmatched.
+- Published kytos/topology.interface.deleted event on interface deletion
 
 [2025.1.1] - 2025-05-06
 ***********************